2.10

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -14,7 +14,7 @@ int _atoi(char *s)
 	int ln = 0, sign = 1, it = 0, i;
 	int order = 1, start = 0, end = 0, s_num = 1;
 	int stop = 1;
-	long num = 0;
+	unsigned int  num = 0;
 
 	if (s[0] == '\0')
 		return (0);
